pyproject.toml
[project]
name = "util"
version = "0.1.0"
description = ""
authors = [
    {name = "Monique Rio",email = "mrio@umich.edu"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer (>=0.21.0,<0.22.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
